requirements.txt
## Web Application
Django==1.8.4
django-autoslug==1.8.0
django-grappelli==2.7.1
django-dotenv==1.3.0
django-model-utils==2.3.1
djangorestframework==3.2.2
django-bootstrap3==6.2.2
django-ipware==1.1.1
whitenoise==2.0.3
gunicorn==19.3.0

## Database Requirements
psycopg2==2.6.1
dj-database-url==0.3.0

## Utilities
python-dateutil==2.4.2
titlecase==0.7.2
six==1.9.0

## Testing
nose==1.3.7
coverage==3.7.1
mock==1.3.0
funcsigs==0.4
pbr==1.5.0
